requirements.txt
flask==2.3.3
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==4.6.0
sqlalchemy==2.0.23
schedule==1.2.0
